requirements.txt
Flask==3.0.0
waitress==2.1.2
Roboflow==1.1.9
python-dotenv==1.0.0